+++ source/extensions/isaacsim.sensors.physx.ui/docs/CHANGELOG.md
@@ -1,4 +1,7 @@
 # Changelog
+## [2.4.2] - 2026-01-24
+### Changed
+- Fix issues with menu click and context menu tests being flaky
 
 ## [2.4.1] - 2026-01-22
 ### Changed